Code Refactoring, final touches. Improving code readability by creating some helper functions and moving them to helper modules. Some small error fixing for tests when loading to BigQuery.
