+++ pipelines/google_sheets/google_sheets.py
@@ -6,9 +6,9 @@ from dlt.common.configuration.specs import GcpClientCredentialsWithDefault
 from dlt.common.typing import DictStrAny, TDataItem
 from dlt.common.exceptions import MissingDependencyException
 import logging
-from pipelines.google_sheets.data_validator import serial_date_to_datetime, get_spreadsheet_id, convert_named_range_to_a1
-from pipelines.google_sheets.api_calls import api_auth
-from pipelines.google_sheets import api_calls
+from pipelines.google_sheets.helpers.data_processing import get_spreadsheet_id, convert_named_range_to_a1, process_range
+from pipelines.google_sheets.helpers.api_calls import api_auth
+from pipelines.google_sheets.helpers import api_calls
 try:
     from apiclient.discovery import build, Resource
 except ImportError:
@@ -43,7 +43,7 @@ def google_spreadsheet(spreadsheet_identifier: str, range_names: list[str] = Non
         if get_named_ranges:
             named_ranges = {convert_named_range_to_a1(named_range_dict=named_range, sheet_names_dict=simple_metadata["sheets"]): named_range["name"] for named_range in simple_metadata["named_ranges"]}
             range_names += list(named_ranges.keys())
-    # get metadata on the first 2 rows for every provided range
+    # get data and metadata
     metadata_ranges_all = api_calls.get_metadata(spreadsheet_id=spreadsheet_id, service=service, ranges=range_names, named_ranges=named_ranges)
     data_resources_list = get_data(service=service, spreadsheet_id=spreadsheet_id, range_names=range_names, metadata_dict=metadata_ranges_all)
     metadata_resource = metadata_table(spreadsheet_info=metadata_ranges_all, spreadsheet_id=spreadsheet_id)
@@ -91,24 +91,9 @@ def get_data(service: Resource, spreadsheet_id: str, range_names: list[str], met
     @:returns: my_resources - list of dlt resources, each containing a table of a specific range
     """
 
-    # handle requests with no ranges - edge case
-    if not range_names:
-        logging.warning("Fetching data error: No ranges to get data from. Check the input ranges are not empty.")
-        return []
+    # get data from Google Sheets and iterate through them to process each range into a separate dlt resource
     my_resources = []
-    # Make an api call to get the data for all sheets and ranges
-    # get dates as serial number
-    values = service.spreadsheets().values().batchGet(
-        spreadsheetId=spreadsheet_id,
-        ranges=range_names,
-        # un formatted returns typed values
-        valueRenderOption="UNFORMATTED_VALUE",
-        # will return formatted dates as a serial number
-        dateTimeRenderOption="SERIAL_NUMBER"
-    ).execute()["valueRanges"]
-    logging.info("Data fetched")
-
-    # iterate through values
+    values = api_calls.get_data_batch(service=service, spreadsheet_id=spreadsheet_id, range_names=range_names)
     for i in range(len(values)):
         # get range name and metadata for sheet
         sheet_range_name = values[i]["range"]
@@ -137,32 +122,3 @@ def get_data(service: Resource, spreadsheet_id: str, range_names: list[str], met
         logging.info(f"Data for {sheet_range_name} loaded")
     logging.info("Finished loading all ranges")
     return my_resources
-
-
-def process_range(sheet_val: Iterator[DictStrAny], sheet_meta: dict[Any]) -> Iterator[DictStrAny]:
-    """
-    Receives 2 arrays of tabular data inside a sheet. This will be processed into a schema that is later stored into a database table
-    @:param: spreadsheet_id - the id of the spreadsheet
-    @:param: spreadsheets_resource - Resource object, used to make calls to google spreadsheets
-    @:return:  table_dict - a dict version of the table. It generates a dict of the type {header:value} for every row.
-    """
-    # get headers and first line data types which is just Datetime or not Datetime so far and loop through the remaining values
-    headers = sheet_meta["headers"]
-    first_line_val_types = sheet_meta["cols_is_datetime"]
-    for row in sheet_val[1:]:
-        table_dict = {}
-        # empty row; skip
-        if not row:
-            continue
-        # process both rows and check for differences to spot dates
-        for val, header, is_datetime in zip(row, headers, first_line_val_types):
-            # 3 main cases: null cell value, datetime value, every other value
-            # handle null values properly. Null cell values are returned as empty strings, this will cause dlt to create new columns and fill them with empty strings
-            if val == "":
-                fill_val = None
-            elif is_datetime:
-                fill_val = serial_date_to_datetime(val)
-            else:
-                fill_val = val
-            table_dict[header] = fill_val
-        yield table_dict
