+++ pipelines/google_sheets/helpers/api_calls.py
@@ -1,12 +1,10 @@
 # Contains helper functions to make API calls
 
-from typing import Any
 import logging
-import dlt
 from dlt.common.configuration.specs import GcpClientCredentialsWithDefault
 from dlt.common.typing import DictStrAny
 from dlt.common.exceptions import MissingDependencyException
-from pipelines.google_sheets.data_validator import get_first_rows, is_date_datatype
+from pipelines.google_sheets.helpers.data_processing import metadata_preprocessing, get_first_line, get_range_headers
 try:
     from apiclient.discovery import build, Resource
 except ImportError:
@@ -24,7 +22,7 @@ def api_auth(credentials: GcpClientCredentialsWithDefault) -> Resource:
     return service
 
 
-def get_metadata_simple(spreadsheet_id: str, service: Resource) -> dict[DictStrAny]:
+def get_metadata_simple(spreadsheet_id: str, service: Resource) -> DictStrAny:
     """
     Makes a simple get metadata API call which just returns information about the spreadsheet such as: sheet_names and named_ranges
     @:param: spreadsheet_id - string containing the id of the spreadsheet
@@ -56,7 +54,7 @@ def get_metadata_simple(spreadsheet_id: str, service: Resource) -> dict[DictStrA
     return return_info
 
 
-def get_metadata(spreadsheet_id: str, service: Resource, ranges: list[str], named_ranges: dict[str] = None) -> dict[Any]:
+def get_metadata(spreadsheet_id: str, service: Resource, ranges: list[str], named_ranges: dict[str] = None) -> DictStrAny:
     """
     # TODO: add fields to save on info returned
     Gets the metadata for the first 2 rows of every range specified. The first row is deduced as the header and the 2nd row specifies the format the rest of the data should follow
@@ -68,37 +66,15 @@ def get_metadata(spreadsheet_id: str, service: Resource, ranges: list[str], name
 
     # process metadata ranges so only the first 2 rows are appended
     # response like dict will contain a dict similar to the response by the Google Sheets API: ranges are returned inside the sheets they belong in the order given in the API request.
-    meta_ranges = []
-    response_like_dict = {}
-    metadata_all_ranges = {}
-    for requested_range in ranges:
-        # convert range to first 2 rows
-        range_info = get_first_rows(requested_range)
-        sheet_name = range_info[0]
-        unfilled_range_dict = {"range": requested_range,
-                               "headers": [],
-                               "cols_is_datetime": [],
-                               "name": None
-                               }
-        # check whether this range is a named range and save info in metadata if so
-        if named_ranges and requested_range in named_ranges:
-            unfilled_range_dict["name"] = named_ranges[requested_range]
-        # add a dict with range info to the list of ranges inside the sheet
-        if sheet_name in response_like_dict:
-            response_like_dict[sheet_name].append(unfilled_range_dict)
-        else:
-            response_like_dict[sheet_name] = [unfilled_range_dict]
-        meta_ranges.append(range_info[1])
-
-    # make call to get metadata
-    # TODO: add fields so the calls are more efficient
+    meta_ranges, response_like_dict = metadata_preprocessing(ranges=ranges, named_ranges=named_ranges)
     spr_meta = service.spreadsheets().get(
         spreadsheetId=spreadsheet_id,
         ranges=meta_ranges,
         includeGridData=True
     ).execute()
 
-    # process and populate metadata in response like dict
+    # process and populate metadata in response like dict but return in from metadata_all_ranges because we need the data returned in a more organized format
+    metadata_all_ranges = {}
     for sheet in spr_meta["sheets"]:
         # get sheet name, so we can associate with dict and load the data into dict
         meta_sheet_name = sheet["properties"]["title"]
@@ -107,63 +83,52 @@ def get_metadata(spreadsheet_id: str, service: Resource, ranges: list[str], name
         # skip record if not found in the response dict
         if not (meta_sheet_name in response_like_dict):
             continue
-
         # get ranges inside the sheet in order
         for i in range(len(sheet_data)):
+            metadata_range_name = response_like_dict[meta_sheet_name][i]["range"]
             # check that sheet is not empty, otherwise delete
             if not ("rowData" in sheet_data[i]):
-                logging.warning(f"Metadata - Skipped empty range: {response_like_dict[meta_sheet_name][i]['range']}")
+                logging.warning(f"Metadata - Skipped empty range: {metadata_range_name}")
                 continue
             # get headers and 1st row data
             range_metadata = sheet_data[i]["rowData"]
-            headers = _get_range_headers(range_metadata=range_metadata)
+            headers = get_range_headers(range_metadata=range_metadata, range_name=metadata_range_name)
             if not headers:
-                logging.warning(f"Metadata: Skipped range with empty headers: {response_like_dict[meta_sheet_name][i]['range']}")
+                logging.warning(f"Metadata: Skipped range with empty headers: {metadata_range_name}")
                 continue
-            first_line_values = _get_first_line(range_metadata=range_metadata)
+            first_line_values = get_first_line(range_metadata=range_metadata)
             if not first_line_values:
-                logging.warning(f"Metadata: No data values for the first line of data {response_like_dict[meta_sheet_name][i]['range']}")
-
+                logging.warning(f"Metadata: No data values for the first line of data {metadata_range_name}")
             # add headers and values
             response_like_dict[meta_sheet_name][i]["headers"] = headers
             response_like_dict[meta_sheet_name][i]["cols_is_datetime"] = first_line_values
-
             # append dict to response
-            metadata_range_name = response_like_dict[meta_sheet_name][i]["range"]
             metadata_all_ranges[metadata_range_name] = response_like_dict[meta_sheet_name][i]
     return metadata_all_ranges
 
 
-def _get_range_headers(range_metadata: dict[Any]) -> list[str]:
+def get_data_batch(service: Resource, spreadsheet_id: str, range_names: list[str]) -> list[DictStrAny]:
     """
-    Helper. Receives the metadata for a range of cells and outputs only the headers for columns, i.e first line of data
-    @:param: range_metadata - Dict containing metadata for the first 2 rows of a range, taken from Google Sheets API response.
-    @:return: headers - list of headers
+    Calls Google Sheets API to get data in a batch. This is the most efficient way to get data for multiple ranges inside a spreadsheet. However, this API call will return the data for each range
+    without the same name that the range was called
+    @:param: service - Object to make api calls to Google Sheets
+    @:param: spredsheet_id - the id of the spreadsheet
+    @:param: range_names - list of range names
+    @:return: values - list of dictionaries, each dictionary will contain all data for one of the requested ranges
     """
-    headers = []
-    empty_header_index = 0
-    for header in range_metadata[0]["values"]:
-        if header:
-            headers.append(header["formattedValue"])
-        else:
-            headers.append(f"empty_header_filler{empty_header_index}")
-            empty_header_index = empty_header_index + 1
-    # manage headers being empty
-    if len(headers) == empty_header_index:
+    # handle requests with no ranges - edge case
+    if not range_names:
+        logging.warning("Fetching data error: No ranges to get data from. Check the input ranges are not empty.")
         return []
-    return headers
-
-
-def _get_first_line(range_metadata: str) -> list[bool]:
-    """
-    Helper. Parses through the metadata for a range and checks whether a column contains datetime types or not
-    @:param: range_metadata - Metadata for first 2 rows in a range
-    @:return: is_datetime_cols - list containing True of False depending on whether the first line of data is a datetime object or not.
-    """
-
-    # get data for 1st column and process them, if empty just return an empty list
-    try:
-        is_datetime_cols = is_date_datatype(range_metadata[1]["values"])
-    except IndexError:
-        return []
-    return is_datetime_cols
+    # Make an api call to get the data for all sheets and ranges
+    # get dates as serial number
+    values = service.spreadsheets().values().batchGet(
+        spreadsheetId=spreadsheet_id,
+        ranges=range_names,
+        # un formatted returns typed values
+        valueRenderOption="UNFORMATTED_VALUE",
+        # will return formatted dates as a serial number
+        dateTimeRenderOption="SERIAL_NUMBER"
+    ).execute()["valueRanges"]
+    logging.info("Data fetched")
+    return values
